Split filter and unfilter files
