+++ man/dust_filter_create.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/interface.R
+% Please edit documentation in R/interface-filter.R
 \name{dust_filter_create}
 \alias{dust_filter_create}
 \title{Create a particle filter}
