+++ man/dust_filter_rng_state.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/interface.R
+% Please edit documentation in R/interface-filter.R
 \name{dust_filter_rng_state}
 \alias{dust_filter_rng_state}
 \alias{dust_filter_set_rng_state}
