+++ man/dust_unfilter_create.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/interface.R
+% Please edit documentation in R/interface-unfilter.R
 \name{dust_unfilter_create}
 \alias{dust_unfilter_create}
 \title{Create an unfilter}
