+++ man/dust_unfilter_last_history.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/interface.R
+% Please edit documentation in R/interface-unfilter.R
 \name{dust_unfilter_last_history}
 \alias{dust_unfilter_last_history}
 \title{Fetch last unfilter history}
